04-AutoGen/01-BloggerAgent/requirement.txt
ipython
pyautogen
matplotlib
yfinance